Pdeathsig is Linux only so split out cases for portability
